Completed all the crud operations

+++ server/database.sql
@@ -3,7 +3,7 @@ CREATE DATABASE transactionsdb;
 CREATE TABLE transactions(
     Id SERIAL PRIMARY KEY,
     TransactionName VARCHAR(255),
-    Ammount MONEY,
+    Ammount INT,
     TransactionDate DATE,
     TransactionType VARCHAR(255)
 );
